+++ packages/lexicon/lightning.js
@@ -3,4 +3,4 @@
  * Note that the major version represents the API spec version, while the minor version
  * represents the lexicon implementation of it
  */
-export const API_VERSION = 1.1;
+export const API_VERSION = 1.2;