+++ packages/ws-worker/src/util/index.ts
@@ -4,6 +4,7 @@ import getWithReply from './get-with-reply';
 import stringify from './stringify';
 import createRunState from './create-run-state';
 import throttle from './throttle';
+export * from './timestamp';
 
 export {
   throttle,