+++ packages/ws-worker/src/util/timestamp.ts
@@ -1,0 +1,2 @@
+export const timeInMicroseconds = (time?: bigint) =>
+  time && (BigInt(time) / BigInt(1e3)).toString();